+++ lib/client/api_service.es6.js
@@ -11,8 +11,6 @@ require('../web_apis/release.es6.js');
 require('../web_apis/release_interface_relationship.es6.js');
 require('../web_apis/web_interface.es6.js');
 require('./api_confluence.es6.js');
-require('./api_matrix.es6.js');
-require('./events.es6.js');
 require('./state.es6.js');
 require('./stats_controller.es6.js');
 const pkg = org.chromium.apis.web;
