+++ lib/controller/api_catalog.es6.js
@@ -6,7 +6,6 @@
 require('angular');
 
 require('../compat.es6.js');
-require('../component/catalog_table.es6.js');
 require('../dao/dao_container.es6.js');
 require('../u2/ActionView.es6.js');
 require('../u2/DAOController.es6.js');
