+++ lib/parse/parser_interpreter.es6.js
@@ -3,7 +3,6 @@
 // found in the LICENSE file.
 'use strict';
 
-require('../dao/grid_dao.es6.js');
 require('../web_apis/api_compat_data.es6.js');
 require('./expressions.es6.js');
