+++ main/app.es6.js
@@ -15,7 +15,6 @@ require('../lib/web_apis/api_compat_data.es6.js');
 require('../lib/web_apis/release.es6.js');
 require('../lib/web_apis/web_interface.es6.js');
 require('../lib/web_apis/release_interface_relationship.es6.js');
-require('../lib/client/api_matrix.es6.js');
 require('../lib/client/api_confluence.es6.js');
 
 let app = angular.module('confluence', ['ui.router']);
