+++ main/relational_to_compat.es6.js
@@ -13,7 +13,6 @@ require('foam2');
 require('../lib/object.es6.js');
 require('../lib/property.es6.js');
 
-require('../lib/dao/grid_dao.es6.js');
 require('../lib/dao/http_json_dao.es6.js');
 require('../lib/dao/json_dao_container.es6.js');
 require('../lib/dao/local_json_dao.es6.js');
