+++ test/any/files-helper.js
@@ -19,8 +19,6 @@ function requireBrowserCode() {
   require('../../lib/chart/time_series_chart.es6.js');
   require('../../lib/client/api_confluence.es6.js');
   require('../../lib/client/api_service.es6.js');
-  require('../../lib/client/release_interface_dao.es6.js');
-  require('../../lib/component/catalog_table.es6.js');
   require('../../lib/controller/api_catalog.es6.js');
   require('../../lib/controller/api_confluence.es6.js');
   require('../../lib/controller/default.es6.js');
@@ -42,9 +40,6 @@ beforeAll(function() {
   if (foam.isServer) requireNodeCode();
   else requireBrowserCode();
 
-  // Client-side, but used in server integration tests.
-  require('../../lib/client/api_matrix.es6.js');
-
   require('../../lib/client/stats_controller.es6.js');
   require('../../lib/compat.es6.js');
   require('../../lib/confluence/aggressive_removal.es6.js');
@@ -59,7 +54,6 @@ beforeAll(function() {
   require('../../lib/confluence/set_ops.es6.js');
   require('../../lib/dao/api_service_dao.es6.js');
   require('../../lib/dao/dao_container.es6.js');
-  require('../../lib/dao/grid_dao.es6.js');
   require('../../lib/dao/http_json_dao.es6.js');
   require('../../lib/dao/indexed_dao.es6.js');
   require('../../lib/dao/json_dao_container.es6.js');
